src/api/requirements.txt
fastapi==0.112.0
pydantic==2.8.2
pydantic-settings==2.4.0
uvicorn==0.30.5
requests==2.32.3
cffi
pillow==11.2.1
tqdm==4.67.1
torch==2.7.0
torchvision==0.22.0
transformers==4.52.4
numpy==1.23.5
faiss-gpu==1.7.2
opencv-python==4.11.0.86
gunicorn==23.0.0